+++ react-app/package.json
@@ -8,6 +8,7 @@
     "@testing-library/user-event": "^12.8.3",
     "http-proxy-middleware": "^1.0.5",
     "react": "^17.0.2",
+    "react-datepicker": "^4.2.0",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.4",
     "react-router-dom": "^5.2.0",
