+++ react-app/public/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <title>Simple React App</title>
+    <title>Rev Up</title>
   </head>
   <body>
     <div id="root"></div>
