+++ react-app/src/components/LoginForm.css
@@ -5,7 +5,6 @@
 	box-sizing: border-box;
     margin: 0;
     padding: 0;
-    background-color: #c0c0c0
 }
 
 li, a, logout-button {
