+++ react-app/src/pages/CreateEvent.js
@@ -1,0 +1,10 @@
+import EventForm from "../components/EventForm"
+
+export default function CreateEvent() {
+    return (
+        <>
+            <h1>Create Event Page</h1>
+            <EventForm />
+        </>
+    )
+}
