+++ react-app/src/pages/events.css
@@ -1,0 +1,72 @@
+.events {
+  display: flex;
+  flex-direction: column;
+  position: relative;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+}
+
+.event__tile--holder {
+  text-align: center;
+  width: 50%;
+}
+
+.event__tile {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  padding-top: 10px;
+  margin: 5px;
+  border-top: 1px solid #8D8D8D;
+}
+
+.event__tile--imageholder {
+  border-radius: .5rem;
+  background-size: cover;
+  height: 125px;
+  width: 175px;
+  overflow: hidden;
+  margin: 5px;
+}
+
+.event__tile--img {
+  position: relative;
+  width: 100%;
+  height: 100%;
+}
+
+.event__tile--adetail {
+  width: 60%;
+  margin: 5px;
+  text-align: left;
+}
+
+.event__tile--adetail > div {
+  margin-bottom: 10px;
+}
+
+.event__detail--start {
+  color: #6A4F32;
+  font-family: darkmode-off, sans-serif;
+  font-style: normal;
+  font-weight: 100;
+}
+
+.event__detail--name {
+  color: #111111;
+  font-family: darkmode-off, sans-serif;
+  font-style: normal;
+  font-weight: 600;
+}
+
+.event__detail--address {
+  color: #464642;
+  font-family: darkmode-off, sans-serif;
+  font-style: normal;
+  font-weight: 300;
+}
+
+.events__datepicker {
+  background-color: #F6F6F6;
+}
